+++ src/tick.h
@@ -4,7 +4,7 @@
 namespace tick {
 unsigned long getTickCount();
 std::chrono::_V2::steady_clock::time_point getCurrent();
-unsigned long getDiff(std::chrono::_V2::steady_clock::time_point start, std::chrono::_V2::steady_clock::time_point end);
+unsigned getDiff(std::chrono::_V2::steady_clock::time_point start, std::chrono::_V2::steady_clock::time_point end);
 };  // namespace tick
 
 #endif  /// of __TICK_H__
